Move `ViewHierarchy` to `maestro` package
